Perform fisher's exact enrichment analysis

Perform fisher's exact enrichment analysis

KO_fisher(
  ko_stat,
  padj_threshold = 0.05,
  logFC_threshold = NULL,
  add_mini = NULL,
  p.adjust.method = "BH",
  type = c("pathway", "module")[1],
  feature = "ko",
  modulelist = NULL,
  verbose = TRUE
)

Arguments

ko_stat

ko_stat dataframe from ko.test.

padj_threshold

p.adjust threshold to determine whether a feature significant or not. p.adjust < padj_threshold, default: 0.05

logFC_threshold

logFC threshold to determine whether a feature significant or not. abs(logFC)>logFC_threshold, default: NULL

add_mini

add_mini when calculate the logFC. e.g (10+0.1)/(0+0.1), default 0.05*min(avg_abundance)

p.adjust.method

The method used for p-value adjustment (default: "BH").

type

"pathway" or "module" for default KOlist_file.

feature

one of "ko", "gene", "compound"

modulelist

NULL or customized modulelist dataframe, must contain "id","K_num","KOs","Description" columns. Take the `KOlist` as example, use custom_modulelist.

verbose

logical

Value

data.frame

See also

Examples

## use `fisher.test` from the `stats` package.
data("reporter_score_res")
fisher_res <- KO_fisher(reporter_score_res)
#> detect the origin_p.adjust, use the origin_p.adjust.
#> ===============================`fisher.test` done===============================