Plot features heatmap

Plot features heatmap

plot_features_heatmap(
  kodf,
  group = NULL,
  metadata = NULL,
  map_id = "map00780",
  select_ko = NULL,
  only_sig = FALSE,
  columns = NULL,
  modulelist = NULL,
  KO_description = FALSE,
  str_width = 50,
  heatmap_param = list()
)

Arguments

kodf

KO_abundance table, rowname is ko id (e.g. K00001),colnames is samples. or result of `get_reporter_score`

group

The compare group (two category) in your data, one column name of metadata when metadata exist or a vector whose length equal to columns number of kodf.

metadata

metadata

map_id

the pathway or module id

select_ko

select which ko

only_sig

only show the significant KOs

columns

change columns

modulelist

NULL or customized modulelist dataframe, must contain "id","K_num","KOs","Description" columns. Take the `KOlist` as example, use custom_modulelist.

KO_description

show KO description rather than KO id.

str_width

str_width to wrap

heatmap_param

parameters pass to pheatmap

Value

ggplot

Examples

# \donttest{
if (requireNamespace("pheatmap")) {
  data("reporter_score_res")
  plot_features_heatmap(reporter_score_res, map_id = "map00780")
}
#> Loading required namespace: pheatmap
# }