Plot a KEGG pathway network

Plot a KEGG pathway network

plot_pathway_net(path_net_c, simplify = FALSE, plot_depth = FALSE, ...)

Arguments

path_net_c

A `metanet` object representing the pathway network.

simplify

Logical, whether to simplify the network by removing multiple edges and loops. Default is `FALSE`.

plot_depth

Logical, whether to plot the network as a tree layout. Default is `FALSE`.

...

Additional arguments passed to `MetaNet::c_net_plot`.

Value

A plot of the pathway network.

Examples

# \donttest{
if (requireNamespace("MetaNet") && requireNamespace("ggkegg")) {
  tmp_dir <- tempdir()
  pcutils::download2("https://rest.kegg.jp/get/ko01521/kgml", file.path(tmp_dir, "ko01521.xml"))
  path_net_c <- c_net_from_pathway_xml(file.path(tmp_dir, "ko01521.xml"))
  plot_pathway_net(path_net_c)
  pathway_net_index(path_net_c)
}
#> Loading required namespace: MetaNet
# }