Reporter score analysis after C-means clustering

Reporter score analysis after C-means clustering

Extract one cluster from rs_by_cm object

Plot c_means result

RSA_by_cm(
  kodf,
  group,
  metadata = NULL,
  k_num = NULL,
  filter_var = 0.7,
  verbose = TRUE,
  method = "pearson",
  ...
)

extract_cluster(rsa_cm_res, cluster = 1)

plot_c_means(
  rsa_cm_res,
  filter_membership,
  mode = 1,
  show.clust.cent = TRUE,
  show_num = TRUE,
  ...
)

Arguments

kodf

KO_abundance table, rowname is ko id (e.g. K00001),colnames is samples.

group

The comparison groups (at least two categories) in your data, one column name of metadata when metadata exist or a vector whose length equal to columns number of kodf. And you can use factor levels to change order.

metadata

sample information data.frame contains group

k_num

if NULL, perform the cm_test_k, else an integer

filter_var

see c_means

verbose

verbose

method

method from reporter_score

...

additional

rsa_cm_res

a cm_res object

cluster

integer

filter_membership

filter membership 0~1.

mode

1~2

show.clust.cent

show cluster center?

show_num

show number of each cluster?

Value

rs_by_cm

reporter_score object

ggplot

See also

Examples

message("The following example require some time to run:")
#> The following example require some time to run:
# \donttest{
if (requireNamespace("e1071") && requireNamespace("factoextra")) {
  data("KO_abundance_test")
  rsa_cm_res <- RSA_by_cm(KO_abundance, "Group2", metadata,
    k_num = 3,
    filter_var = 0.7, method = "pearson", perm = 199
  )
  extract_cluster(rsa_cm_res, cluster = 1)
}
#> Loading required namespace: e1071
# }