plot_KEGG_map

plot_KEGG_map

plot_KEGG_map(
  ko_stat,
  map_id = "map00780",
  modulelist = NULL,
  type = "pathway",
  feature = "ko",
  color_var = "Z_score",
  save_dir,
  color = c("seagreen", "grey", "orange")
)

Arguments

ko_stat

ko_stat result from pvalue2zs or result of `get_reporter_score`

map_id

the pathway or module id

modulelist

NULL or customized modulelist dataframe, must contain "id","K_num","KOs","Description" columns. Take the `KOlist` as example, use custom_modulelist.

type

"pathway" or "module" for default KOlist for microbiome, "CC", "MF", "BP", "ALL" for default GOlist for homo sapiens. And org in listed in 'https://www.genome.jp/kegg/catalog/org_list.html' such as "hsa" (if your kodf is come from a specific organism, you should specify type here).

feature

one of "ko", "gene", "compound"

color_var

use which variable to color

save_dir

where to save the png files

color

color

Value

png files

References

https://zhuanlan.zhihu.com/p/357687076

Examples

message("The following example will download some files, run yourself:")
#> The following example will download some files, run yourself:
# \donttest{
if (requireNamespace("pathview")) {
  output_dir <- tempdir()
  data("reporter_score_res")
  plot_KEGG_map(reporter_score_res$ko_stat,
    map_id = "map00780", type = "pathway",
    feature = "ko", color_var = "Z_score", save_dir = output_dir
  )
}
#> Loading required namespace: pathview
# }